Rename Python package shaprpy -> pyshapr (canonical code)
Move python/src/shaprpy/ to python/src/pyshapr/ and update internal
imports, docstrings, and error text. Bump distribution name to pyshapr
and version to 0.5.0; update pyproject packaging, MANIFEST.in, and
regenerate uv.lock.
