+++ python/MANIFEST.in
@@ -1,3 +1,3 @@
 include LICENSE
 include README.md
-recursive-include src/shaprpy *.py
+recursive-include src/pyshapr *.py
