+++ python/src/pyshapr/_rutils.py
