+++ python/src/pyshapr/utils.py
